requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
branca==0.7.2
build==1.2.1
build-flask-app==0.1.0
cached-property==1.5.2
cachelib==0.9.0
cargo==0.3
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
Choco==1.0.5
click==8.1.7
cloudpathlib==0.18.1
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
executing==2.0.1
fastjsonschema==2.20.0
fitfile==1.1.7
fitparse==1.2.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
fqdn==1.5.1
future==1.0.0
garminconnect==0.2.23
garmindb==3.6.0
garth==0.4.46
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idbutils==1.1.0
idna==3.8
iniconfig==2.0.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
leaflet==0.0.3
lxml==5.3.0
make==0.1.6.post2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandarallel==1.6.5
pandas==2.2.2
pandastable==0.13.1
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyInquirer==1.0.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pymongo==4.8.0
pymssql==2.3.1
pyodbc==5.1.0
pyparsing==3.1.4
pyproject_hooks==1.1.0
PySimpleGUI==5.0.6
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.3
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.25.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.19.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==74.0.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sklearn-preprocessing==0.1.0
smart-open==7.0.4
snakemd==2.2.0
sniffio==1.3.1
soupsieve==2.5
source==1.2.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
stack-data==0.6.3
tcxfile==1.0.4
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
ttkbootstrap==1.10.1
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
waitress==3.0.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
widgetsnbextension==4.0.11
withings-sync==4.2.5
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xlwings==0.32.2
xyzservices==2024.6.0
zope.event==5.0
zope.interface==7.0.1